Delete historical generation code
